+++ packages/featureserver/src/query.js
@@ -33,13 +33,15 @@ function query (data, params = {}) {
   const queriedData = filtersApplied.all ? data : Winnow.query(data, options)
 
   // ArcGIS client warnings
-  if (options.toEsri) {
-    if (!hasIdField) console.warn(`The requested provider has no "idField" assignment. This can cause errors in ArcGIS clients`)
-    else if (data.metadata.idField.toLowerCase() === 'objectid' && data.metadata.idField !== 'OBJECTID') {
-      console.warn(`The requested provider's "idField" is a mixed-case version of "OBJECTID". This can cause errors in ArcGIS clients`)
-    } else if (queriedData.features.some(feature => { return !Number.isInteger(feature.attributes.OBJECTID) || feature.attributes.isArrayOBJECTID > 2147483647 })) {
-      console.warn(`OBJECTIDs created from provider's "idField" are not integers from 0 to 2147483647`)
-    }
+  if (options.toEsri && !hasIdField) {
+    console.warn(`WARNING: requested provider has no "idField" assignment. This can cause errors in ArcGIS clients`)
+  } else if (options.toEsri && data.metadata.idField.toLowerCase() === 'objectid' && data.metadata.idField !== 'OBJECTID') {
+    console.warn(`WARNING: requested provider's "idField" is a mixed-case version of "OBJECTID". This can cause errors in ArcGIS clients`)
+  }
+
+  // Compare provider metadata fields to feature properties
+  if (_.has(data, 'metadata.fields') && _.has(data, 'features[0].properties')) {
+    warnOnMetadataFieldDiscrepencies(data.metadata.fields, data.features[0].properties)
   }
 
   if (params.f === 'geojson') return { type: 'FeatureCollection', features: queriedData.features }
@@ -104,4 +106,33 @@ function queryStatistics (data, params) {
     return { attributes: row }
   })
   return renderStatistics(statResponse, params)
+}
+
+/**
+ * Compare fields generated from metadata to properties of a data feature.
+ * Warn if differences discovered
+ * @param {*} metadataFields
+ * @param {*} properties
+ */
+function warnOnMetadataFieldDiscrepencies (metadataFields, featureProperties) {
+  // build a comparison collection from the data samples properties
+  const featureFields = Object.keys(featureProperties).map(name => {
+    return { name, type: Utils.detectType(featureProperties[name]) }
+  })
+
+  // compare metadata to feature properties; identifies fields defined in metadata that are not found in feature properties
+  // that have a metadata type definition inconsistent with feature property's value
+  metadataFields.filter(field => {
+    let featureField = _.find(featureFields, ['name', field.name])
+    if (!featureField || (field.type !== featureField.type && !(field.type === 'Date' && featureField.type === 'Integer') && !(field.type === 'Double' && featureField.type === 'Integer'))) {
+      console.warn(`WARNING: requested provider's metadata field "${field.name} (${field.type})" not found in feature properties)`)
+    }
+  })
+
+  // compare feature properties to metadata fields; identifies fields found on feature that are not defined in metadata field array
+  featureFields.filter(field => {
+    if (!_.find(metadataFields, ['name', field.name])) {
+      console.warn(`WARNING: requested provider's features have property "${field.name} (${field.type})" that was not defined in metadata fields array)`)
+    }
+  })
 }